Started new version of Dylan blockout based on class feedback
Started new Blockout_DylanRedone. It has a bigger layout since the teacher said that was okay. I also used the landscaping tool to make the new terrain. The general layout looks a lot closer to the original concept art now. I've only done new landscaping and moved the lighthouse, I still need to redo the beginning walkways and the cave